- fixed error in solver to update scope_vars after new step size - added possibility of controlling tolerance in RK45 method - fixed problem where scope_vars were not re-rolled after initial step size (this can be fixed more elegantly). - fixed tests relative imports

+++ numerous/engine/model/numba_model.py
@@ -100,7 +100,7 @@ class NumbaModel:
 
         return np.array(result, dtype=np.float64)
 
-    def historian_update(self, time: int) -> None:
+    def historian_update(self, time: np.float64) -> None:
         ix = self.historian_ix
         varix = 1
         self.historian_data[0][ix] = time
